requirements.txt
opencv-python
scipy>=0.17.0
scikit-image
numpy
matplotlib
Pillow>=4.3.0
imgaug
tensorflow
git+https://github.com/lanpa/tensorboardX
joblib
torch>=1.3.0
torchvision>=0.4.1
